# openapi.yaml
openapi: 3.0.3
info:
  title: Library Management System API
  version: 1.0.0
  description: API for managing books, members, loans, copies, fines and categories.
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /members:
    get:
      tags: [Members]
      summary: List members (Admin)
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      tags: [Members]
      summary: Add member (Admin)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberCreate"
      responses:
        "201":
          description: Created
  /members/{memberId}:
    get:
      tags: [Members]
      summary: Get member
      parameters:
        - $ref: "#/components/parameters/memberId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    put:
      tags: [Members]
      summary: Update member (Admin)
      parameters:
        - $ref: "#/components/parameters/memberId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberCreate"
      responses:
        "200":
          description: Updated
    delete:
      tags: [Members]
      summary: Delete member (Admin)
      parameters:
        - $ref: "#/components/parameters/memberId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    memberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    Member:
      type: object
      properties:
        member_id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        role:
          type: string
    MemberCreate:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        password:
          type: string
        role:
          type: string
...